Move matrix package inside kodr_internals package
