+++ full/decoder.go
@@ -4,7 +4,7 @@ import (
 	"github.com/cloud9-tools/go-galoisfield"
 	"github.com/itzmeanjan/kodr"
 	"github.com/itzmeanjan/kodr/kodr_internals"
-	"github.com/itzmeanjan/kodr/matrix"
+	"github.com/itzmeanjan/kodr/kodr_internals/matrix"
 )
 
 type FullRLNCDecoder struct {
