+++ full/recoder.go
@@ -3,7 +3,7 @@ package full
 import (
 	"github.com/cloud9-tools/go-galoisfield"
 	"github.com/itzmeanjan/kodr/kodr_internals"
-	"github.com/itzmeanjan/kodr/matrix"
+	"github.com/itzmeanjan/kodr/kodr_internals/matrix"
 )
 
 type FullRLNCRecoder struct {
