+++ kodr_internals/matrix/decoder_state.go
